+++ src/main/java/com/bookbla/americano/domain/notification/service/MailService.java
@@ -1,4 +1,4 @@
-package com.bookbla.americano.domain.member.service;
+package com.bookbla.americano.domain.notification.service;
 
 public interface MailService {
     void sendTransactionFailureEmail(String transactionName, String message);
