+++ src/main/java/com/bookbla/americano/domain/schedule/SchedulerService.java
@@ -6,7 +6,7 @@ import com.bookbla.americano.base.log.discord.BookblaLogDiscord;
 import com.bookbla.americano.domain.member.repository.MemberEmailRepository;
 import com.bookbla.americano.domain.member.repository.MemberPostcardRepository;
 import com.bookbla.americano.domain.member.repository.MemberRepository;
-import com.bookbla.americano.domain.member.service.MailService;
+import com.bookbla.americano.domain.notification.service.MailService;
 import com.bookbla.americano.domain.schedule.dto.AdminLogEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
